+++ CHANGELOG.md
@@ -7,3 +7,6 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 - Initial release
+- Added World creation
+- Added Campaign creation
+- Added Adventure creation